Make Number parameters and signals into Bytes for speed and to limit high values

+++ src/main/java/dev/hyperlynx/pulsetech/core/protocol/ExecutionContext.java
@@ -2,4 +2,4 @@ package dev.hyperlynx.pulsetech.core.protocol;
 
 import java.util.List;
 
-public record ExecutionContext(ProtocolBlockEntity block, List<Short> params) {}
+public record ExecutionContext(ProtocolBlockEntity block, List<Byte> params) {}
