+++ src/main/java/dev/hyperlynx/pulsetech/feature/number/NumberKnower.java
@@ -1,5 +1,5 @@
 package dev.hyperlynx.pulsetech.feature.number;
 
 public interface NumberKnower {
-    short getNumber();
+    byte getNumber();
 }
